Cleaned up mangaer/CLI
There needs to be some engine bindings, those are marked by todos.
There is currently a mock engine that you can connect to and play with

The timing may also need some work its not perfect. I don't know if
perf_counter is what we want but we can play around with that.